* p4a-pips:
  Fix bug in recover_for_loop newly created entities were not added to module
  fix some warnings with gcc-4.4.1 that prevent compilation with -O0 -g -Wall -Werror (autotools --enable-devel-mode option)
  Improve partial_eval to handle neutral element generated by partial_eval_biary_operator improvements show in Expressions/partial_eval0[12].tpips
  forgotten references.
  update effects.tex documentation. Add GAPs domains.

+++ packages/PIPS/pips/src/Passes/gpips/gtk_status.c
@@ -90,7 +90,7 @@ void display_memory_usage() {
 			"_text %#x, _data %#x, _end %#x, _brk %#x\n", &etext, &edata, &end,
 			sbrk(0));
 
-	sprintf(memory_string, "%10.3f", ((int) sbrk(0) - (int) &etext)
+	sprintf(memory_string, "%10.3f", ((intptr_t) sbrk(0) - (intptr_t) &etext)
 			/ (double) (1 << 20));
 
 	gtk_label_set_text(GTK_LABEL(memory_name), memory_string);
